+++ pybop/plot/trajectories.py
@@ -1,4 +1,4 @@
-from pybop.plot.standard_plots import StandardPlot
+from pybop.plot.util import get_backend_from_figure, parse_data, wrap_text
 
 
 def trajectories(
@@ -49,21 +49,35 @@ def trajectories(
     plotly.graph_objs.Figure
         The Plotly figure object for the scatter plot.
     """
-    plot_dict = StandardPlot(
-        x,
-        y,
-        title=title,
-        xaxis_title=xaxis_title,
-        yaxis_title=yaxis_title,
-        labels=labels,
-        label_width=label_width,
-        style={"height": 600, "width": 600, "bg_color": "white"},
-        legend_style={},
-        backend=backend,
-        figures=figures,
-        axes=axes,
+    backend = get_backend_from_figure(backend, figures)
+    figures, axes, create_figure, _ = backend.parse_input_axes(
+        figures, axes, num_plots=1
     )
 
-    fig = plot_dict(show=show)
+    if create_figure:
+        fig = backend.create_figure(
+            style={"height": 600, "width": 600, "bg_color": "white"}
+        )
+    else:
+        fig = figures[0]
 
-    return fig
+    backend.update_axes_titles(fig, axes[0], xaxis_title, yaxis_title)
+    backend.update_plot_titles(fig, axes[0], title)
+
+    x, y = parse_data(x, y)
+    xi = x[0]
+    for i in range(0, len(y)):
+        if len(x) > 1:
+            xi = x[i]
+        label = None
+        if labels is not None:
+            label = wrap_text(labels[i], label_width, backend=backend.name)
+
+        backend.plot_trace(backend.line(xi, y[i], label), fig, ax=axes[0])
+
+    backend.legend(fig, axes=axes[0])
+
+    if show:
+        backend.show_figure(fig)
+    else:
+        return fig